fixed meal log, issue was float values and adapter

+++ app/src/main/java/com/fitwizard/fitwizard/FoodData.java
@@ -1,12 +1,10 @@
 package com.fitwizard.fitwizard;
 
 
-
 // Formating and information to be pulled
 //Used by FoodSearchActivity
 //Used by FoodAdapter for getView()
 public class FoodData {
-
     private final String name;
     private final int calories;
     private final float protein;
@@ -14,8 +12,8 @@ public class FoodData {
     private final float carbs;
     private final String servingSize;
 
-    //Food info to get
-    public FoodData(String name, int calories, int protein, int fats, int carbs, String servingSize) {
+    
+    public FoodData(String name, int calories, float protein, float fats, float carbs, String servingSize) {
         this.name = name;
         this.calories = calories;
         this.protein = protein;
@@ -25,10 +23,9 @@ public class FoodData {
     }
 
     public String getName() { return name; }
-    public int getCalories() { return calories; }  // keep as an INT value
+    public int getCalories() { return calories; }
     public float getProtein_g() { return protein; }
     public float getFats_g() { return fats; }
     public float getCarbs_g() { return carbs; }
     public String getServingSize() { return servingSize; }
 }
-
